requirements.txt
requests>=2.31.0
feedparser>=6.0.11
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.2.2
Flask>=3.0.3
